Feat: category
기능 보완

작성자: 정효주

+++ anymall/main/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import CustomUser, Product  # 이렇게 변경합니다.
+from django.http import JsonResponse
+from .models import *
 
 def home(request):
     # user = CustomUser.objects.get()  # User 대신 CustomUser를 사용합니다.
@@ -80,8 +81,6 @@ def mypage(request):
 def admin_set(request):
     return render(request, "admin_set.html")
 
-from .models import *
-
 def admin_category(request):
     if request.method == "POST":
         category_name = request.POST.get("category_name")
@@ -92,10 +91,20 @@ def admin_category(request):
         categories = Category.objects.all()
         return render(request, "admin_category.html", {"categories": categories})
 
+from django.shortcuts import render, redirect
+from .models import Category, Product
+
 def delete_category(request, category_id):
+    template = "admin_category.html"
     if request.method == "POST":
         category = Category.objects.get(pk=category_id)
-        category.delete()
-        return redirect("admin_category")
+
+        # 카테고리가 Product 모델에서 사용 중인지 확인
+        if Product.objects.filter(category_id=category_id).exists():
+            context = {"result": "이 카테고리는 하나 이상의 제품과 연관되어 있어 삭제할 수 없습니다."}
+            return render(request, template, context)
+        else:
+            category.delete()
+            return redirect("admin_category")
     else:
-        return render(request, "admin_category.html")
+        return render(request, template, context)
